+++ parser/src/ast/attribute.rs
@@ -13,11 +13,18 @@ pub struct Attribute {
 impl Attribute {
     pub(crate) fn parse(pair: Pair<Rule>) -> Self {
         assert_eq!(pair.as_rule(), Rule::attribute);
+        Self::parse_impl(pair)
+    }
 
+    pub(crate) fn parse_inline(pair: Pair<Rule>) -> Self {
+        assert_eq!(pair.as_rule(), Rule::attribute_inline);
+        Self::parse_impl(pair)
+    }
+
+    fn parse_impl(pair: Pair<Rule>) -> Self {
         let span = Span::from_pair(&pair);
 
         let mut pairs = pair.into_inner();
-        pairs.next().unwrap(); // Skip #.
         pairs.next().unwrap(); // Skip [.
 
         let pair = pairs.next().unwrap();
@@ -32,10 +39,12 @@ impl Attribute {
         for pair in pairs {
             match pair.as_rule() {
                 Rule::ident => att.options.push(Ident::parse(pair)),
+
                 Rule::tok_par_open
                 | Rule::tok_par_close
                 | Rule::tok_comma
                 | Rule::tok_squ_close => {}
+
                 _ => unreachable!(),
             }
         }