+++ parser/src/ast/prelude.rs
@@ -7,6 +7,7 @@ pub(crate) struct Prelude {
     doc: DocString,
     doc_inline: DocString,
     attrs: Vec<Attribute>,
+    attrs_inline: Vec<Attribute>,
 }
 
 impl Prelude {
@@ -14,12 +15,18 @@ impl Prelude {
         let mut doc = DocString::new();
         let mut doc_inline = DocString::new();
         let mut attrs = Vec::new();
+        let mut attrs_inline = Vec::new();
 
         while let Some(pair) = pairs.peek() {
             match pair.as_rule() {
                 Rule::doc_string if !inline => doc.push(pair),
                 Rule::doc_string_inline if inline => doc_inline.push_inline(pair),
                 Rule::attribute if !inline => attrs.push(Attribute::parse(pair)),
+
+                Rule::attribute_inline if inline => {
+                    attrs_inline.push(Attribute::parse_inline(pair))
+                }
+
                 _ => break,
             }
 
@@ -30,6 +37,7 @@ impl Prelude {
             doc,
             doc_inline,
             attrs,
+            attrs_inline,
         }
     }
 
@@ -43,5 +51,9 @@ impl Prelude {
 
     pub(crate) fn take_attrs(&mut self) -> Vec<Attribute> {
         mem::take(&mut self.attrs)
+    }
+
+    pub(crate) fn take_attrs_inline(&mut self) -> Vec<Attribute> {
+        mem::take(&mut self.attrs_inline)
     }
 }